+++ src/generator/sphere.hpp
@@ -1,14 +1,15 @@
 #ifndef SPHERE_H
 #define SPHERE_H
 #include "../utils/point.hpp"
+#include "../utils/types.hpp"
 
 #include <vector>
 
 class Sphere {
 private:
     float _radius;
-    float _slices;
-    float _stacks;
+    u32 _slices;
+    u32 _stacks;
 
 public:
     Sphere(int argc, char** argv);
